Update README.md
Added some clarifying wordage to the itch.io version not being tested and a message stating to get ahold of me in the discord if someone can confirm it works if it ever happens. That way, I can modify the json and this Readme again to state they both work interchangeably.